# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: CafeOre POS API
  version: 1.0.0
  description: 珈琲・俺POSシステムのAPI仕様
servers:
  - url: http://localhost:8080
    description: 開発環境
paths:
...
  /api/color-settings:
    get:
      summary: 背景色設定一覧取得
      operationId: getColorSettings
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ColorSettingResponse'
    put:
      summary: 背景色設定の作成・更新
      description: 対象（target_type, target_id）と画面（screen）の組が既にあれば色を上書きし、無ければ作成する
      operationId: upsertColorSetting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ColorSettingUpsertRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ColorSettingResponse'
        '400':
          description: リクエストが不正
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/color-settings/{id}:
    delete:
      summary: 背景色設定削除
      operationId: deleteColorSetting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
        '404':
          description: 背景色設定が見つかりません
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/orders:
    get:
      summary: オーダー一覧取得
      operationId: getOrders
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  $ref: '#/components/schemas/OrderResponse'
    post:
      summary: オーダー作成
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /api/orders/{id}:
    get:
      summary: idからオーダー情報取得
      operationId: getOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    put:
      summary: オーダー情報更新
      operationId: updateOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdateRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    delete:
      summary: オーダー削除
      operationId: deleteOrder
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 成功
...
components:
  schemas:
...
    ItemResponse:
      type: object
      required:
        - id
        - name
        - abbr
        - item_type
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbr:
          type: string
        item_type:
          $ref: '#/components/schemas/ItemTypeResponse'
    # 作成リクエスト用（IDや自動生成フィールドを除外）
...
    MenuItemResponse:
      type: object
      required:
        - item
        - quantity
      properties:
        item:
          $ref: '#/components/schemas/ItemResponse'
        quantity:
          type: integer
          minimum: 1
...
    MenuResponse:
      type: object
      required:
        - id
        - name
        - abbr
        - price
        - key
        - items
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        abbr:
          type: string
        price:
          type: integer
        key:
          type: string
        items:
          type: array
          items:
            $ref: '#/components/schemas/MenuItemResponse'
...
    ItemTypeResponse:
      type: object
      required:
        - id
        - name
        - display_name
        - makes_cup
        - needs_brew
        - senior_only
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
        makes_cup:
          type: boolean
          description: この種類のアイテムは1杯ずつカップを作る（注文のカップ・マスター・提供画面に出る）。グッズは false
        needs_brew:
          type: boolean
          description: この種類のアイテムは抽出が要る（割引の対象の杯数・ドリッパーの割り振りに数える）。makes_cup が false なら必ず false
        senior_only:
          type: boolean
          description: この種類のアイテムは上級生だけが淹れる（限定）。needs_brew が false なら必ず false
...
    MenuInfo:
      type: object
      required:
        - id
        - menu_name
        - unit_price
        - menu
        - assignee
        - dripper
      properties:
        id:
          type: string
          format: uuid
          description: 注文明細ID
        menu_name:
          type: string
          description: 注文時点のメニュー名
        unit_price:
          type: integer
          description: 注文時点のメニュー価格
        menu:
          $ref: '#/components/schemas/MenuResponse'
        assignee:
          type: string
          nullable: true
          description: 指名の自由記述（ラベルに印刷する文）。dripper が無い明細では null。番号より前の注文は自由記述だけのことがある
        dripper:
          type: integer
          minimum: 1
          maximum: 6
          nullable: true
          description: 指名したドリッパーの番号（1st〜6th は 1〜6）。指名しない明細は null
    # 注文の1杯。注文明細（メニュー）の構成品を数量分に展開したもの（グッズは含まない）
    OrderCupResponse:
      type: object
      required:
        - id
        - order_menu_id
        - item
        - ready_at
        - served_at
        - dripper
        - dripper_position
        - drip_id
        - brew_started_at
        - brew_finished_at
        - emergency_at
        - emergency_drip_id
        - emergency_printed_at
      properties:
        id:
          type: string
          format: uuid
        order_menu_id:
          type: string
          format: uuid
          description: このカップを含む注文明細のID（MenuInfo.id）
        item:
          $ref: '#/components/schemas/ItemResponse'
        ready_at:
          type: string
          format: date-time
          nullable: true
          description: このカップが準備完了になった時刻。未準備なら null
        served_at:
          type: string
          format: date-time
          nullable: true
          description: このカップを提供した時刻。未提供なら null
        dripper:
          type: integer
          nullable: true
          minimum: 1
          maximum: 6
          description: CaOS が置いたドリッパーの番号（1〜6）。未割当なら null。以下の CaOS の列は CaOS 以外の画面は読まない
        dripper_position:
          type: number
          format: double
          nullable: true
          description: CaOS のドリッパーの中の順番（小さいほど先）
        drip_id:
          type: string
          format: uuid
          nullable: true
          description: CaOS のカードの印。同じ値のカップを 1 枚のカード（1 回のドリップ）で淹れる
        brew_started_at:
          type: string
          format: date-time
          nullable: true
          description: CaOS で抽出を始めた時刻
        brew_finished_at:
          type: string
          format: date-time
          nullable: true
          description: CaOS で抽出を終えた時刻
        emergency_at:
          type: string
          format: date-time
          nullable: true
          description: |
            緊急（入れ直し）にした時刻。緊急でなければ null（POST /api/caos/emergency で付ける。2 回は付けない）。
            緊急にしたとき、dripper・dripper_position・brew_started_at・brew_finished_at は空に戻り、入れ直しのカードの値になる（drip_id は最初に淹れたカードのまま）
        emergency_drip_id:
          type: string
          format: uuid
          nullable: true
          description: 入れ直しで淹れるカードの印。緊急のカップでは CaOS のカードはこちらで決まる。null なら未割当の緊急のカード
        emergency_printed_at:
          type: string
          format: date-time
          nullable: true
          description: |
            緊急のシールを印刷した時刻。emergency_at があってこれが null のカップは、プリンターにつないだレジが印刷する
            （POST .../emergency-label/claim で付けられたときだけ印刷する。失敗したら .../release で null に戻す）
    # Create用のMenuInfo
    MenuInfoCreate:
      type: object
      required:
        - menu_id
        - assignee
        - dripper
      properties:
        order_menu_id:
          type: string
          format: uuid
          description: 更新時に残す既存明細のID。新規明細では省略する。価格・名称はサーバーが保存する。
        menu_id:
          type: string
          format: uuid
        assignee:
          type: string
          nullable: true
          description: 指名の自由記述。新しい明細では dripper が無いと付けられない。空白だけなら null として扱う
        dripper:
          type: integer
          minimum: 1
          maximum: 6
          nullable: true
          description: 指名したドリッパーの番号（1〜6）。指名しない明細は null
    # レスポンス用（完全な情報）
    OrderResponse:
      type: object
      required:
        - id
        - order_id
        - created_at
        - billing_amount
        - received
        - menus
        - cups
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: integer
        created_at:
          type: string
          format: date-time
        ready_at:
          type: string
          format: date-time
          nullable: true
        served_at:
          type: string
          format: date-time
          nullable: true
        billing_amount:
          type: integer
        received:
          type: integer
        discount_order_id:
          type: integer
          nullable: true
        discount_order_cups:
          type: integer
        menus:
          type: array
          items:
            $ref: '#/components/schemas/MenuInfo'
        cups:
          type: array
          description: 注文のカップ（1杯ずつ）。注文した順に並ぶ。グッズだけの注文では空
          items:
            $ref: '#/components/schemas/OrderCupResponse'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentResponse'

    # 作成リクエスト用（IDや自動生成フィールドを除外）
    OrderCreateRequest:
      type: object
      required:
        - order_id
        - billing_amount
        - received
        - menu_ids
      properties:
        order_id:
          type: integer
          example: 1
        billing_amount:
          type: integer
          example: 500
        received:
          type: integer
          example: 500
          default: 0
        discount_order_id:
          type: integer
          nullable: true
        discount_order_cups:
          type: integer
          default: 0
        menu_ids:
          type: array
          items:
            $ref: '#/components/schemas/MenuInfoCreate'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/CommentCreateRequest'

    # 更新リクエスト用
    OrderUpdateRequest:
      type: object
      required:
        - id
        - order_id
        - billing_amount
        - received
        - menu_ids
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: integer
        ready_at:
          type: string
          format: date-time
          nullable: true
        served_at:
          type: string
          format: date-time
          nullable: true
        billing_amount:
          type: integer
        received:
          type: integer
        discount_order_id:
          type: integer
          nullable: true
        discount_order_cups:
          type: integer
        menu_ids:
          type: array
          items:
            $ref: '#/components/schemas/MenuInfoCreate'

    # 一覧取得用（軽量版、関連データを含まない場合）
...
    CommentCreateRequest:
      type: object
      required:
        - author
        - text
      properties:
        author:
          type: string
        text:
          type: string
    CommentResponse:
      type: object
      required:
        - order_id
        - author
        - text
        - created_at
      properties:
        order_id:
          type: string
          format: uuid
        author:
          type: string
        text:
          type: string
        created_at:
          type: string
          format: date-time
...
    ColorTargetType:
      type: string
      description: 背景色を設定する対象の種類
      enum:
        - Item
        - ItemType
      # 既定の定数名（Item, ItemType）は DB モデルの構造体名と衝突するため明示する
      x-enum-varnames:
        - ColorTargetTypeItem
        - ColorTargetTypeItemType
    ColorScreen:
      type: string
      description: |
        背景色を適用する画面。
        cashier はレジのメニューのボタン、cashier_order はレジの過去の注文のカード、
        master・serve はマスター・提供画面のカップ
      enum:
        - cashier
        - cashier_order
        - master
        - serve
      x-enum-varnames:
        - ColorScreenCashier
        - ColorScreenCashierOrder
        - ColorScreenMaster
        - ColorScreenServe
    ColorSettingResponse:
      type: object
      required:
        - id
        - target_type
        - target_id
        - screen
        - color
      properties:
        id:
          type: string
          format: uuid
        target_type:
          $ref: '#/components/schemas/ColorTargetType'
        target_id:
          type: string
          format: uuid
          description: Item または ItemType の ID
        screen:
          $ref: '#/components/schemas/ColorScreen'
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          example: "#bfdbfe"
    ColorSettingUpsertRequest:
      type: object
      required:
        - target_type
        - target_id
        - screen
        - color
      properties:
        target_type:
          $ref: '#/components/schemas/ColorTargetType'
        target_id:
          type: string
          format: uuid
          description: Item または ItemType の ID
        screen:
          $ref: '#/components/schemas/ColorScreen'
        color:
          type: string
          pattern: '^#[0-9a-fA-F]{6}$'
          example: "#bfdbfe"
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: "Invalid order ID format"
...